requirements.txt
playwright==1.47.0
pytest
pytest-xdist
allure-pytest
